object creation methods is now automated in properties_generator.rb
