+++ atome/lib/atome.rb
@@ -39,18 +39,18 @@ require "atome/renderers/html/properties/material"
 require "atome/renderers/html/helpers/html_helpers"
 require "atome/renderers/html"
 # generated property
-require "atome/generated_properties/communications"
-require "atome/generated_properties/effects"
-require "atome/generated_properties/events"
-require "atome/generated_properties/geometries"
-require "atome/generated_properties/helpers"
-require "atome/generated_properties/hierarchies"
-require "atome/generated_properties/identities"
-require "atome/generated_properties/medias"
-require "atome/generated_properties/inputs"
-require "atome/generated_properties/spatials"
-require "atome/generated_properties/utilities"
-require "atome/generated_properties/materials"
+require "atome/generated_methods/communications"
+require "atome/generated_methods/effects"
+require "atome/generated_methods/events"
+require "atome/generated_methods/geometries"
+require "atome/generated_methods/helpers"
+require "atome/generated_methods/hierarchies"
+require "atome/generated_methods/identities"
+require "atome/generated_methods/medias"
+require "atome/generated_methods/inputs"
+require "atome/generated_methods/spatials"
+require "atome/generated_methods/utilities"
+require "atome/generated_methods/materials"
 # property processors
 require "atome/processors/communication"
 require "atome/processors/effect"
@@ -66,9 +66,9 @@ require "atome/processors/material"
 # helpers
 require "atome/helpers/atome_helpers"
 require "atome/helpers/internal_helpers"
-require "atome/generated_properties/atome_methods"
+require "atome/generated_methods/atome_methods"
 # main atome builder
-require "atome/generated_properties/batch"
+require "atome/generated_methods/batch"
 require "atome/builder/properties"
 #Quark.genesis
 require "atome/builder/atome_generator"
@@ -79,4 +79,5 @@ require "atome/environment/device"
 # elementary atomes for basic environment
 require "atome/environment/initialize"
 # methods to simplify object creation
-require "atome/extensions/atome_object_creator"
+require "atome/generated_methods/atome_object_creator"
+require "atome/extensions/atome_code_creator"
