+++ atome/lib/atome/environment/device.rb
@@ -5,7 +5,7 @@ class Device
     # the object below is used to store the basic presets for common type of atomes
     Atome.new({ atome_id: :preset, id: :preset, content: default_values, render: false })
     # the object below is used to hide some atomes rom the device view while still rendered
-    Atome.new({ atome_id: :black_hole, id: :black_hole, content: {}, width: 0, height: 0, x: 0,y: 0, render: false })
+    Atome.new({ atome_id: :black_hole, id: :black_hole, content: {}, width: 0, height: 0, x: 0, y: 0, render: false })
     # the object below is used to store the preferences and settings for current device
     Atome.new({ atome_id: :device, id: :device, language: :english, type: :particle, width: "100%", height: "100%", x: 0, y: 0, z: 1, overflow: :hidden, color: :transparent })
     # the object below hold all the tool that modify atomes
