+++ atome/lib/atome/renderers/html/properties/inputs.rb
@@ -1,7 +1,14 @@
 def camera_html(value="")
+  alert "camera ok!"
   value
 end
 
 def microphone_html(value="")
+  alert "microphone ok!"
+  value
+end
+
+def midi_html(value="")
+  alert "midi ok!"
   value
 end
